sardis-core/pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "sardis-core"
version = "0.1.0"
description = "Core AP2/TAP-aligned domain primitives for the Sardis stablecoin execution layer"
authors = [{name = "Sardis"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.6",
    "pydantic-settings>=2.0",
    "PyNaCl>=1.5",
    "cryptography>=41.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/sardis_v2_core"]
